+++ simconnect-sdk/src/domain/mod.rs
@@ -1,15 +1,15 @@
+mod client_event;
 mod condition;
 mod data_type;
-mod events;
 mod facilities;
 mod notification;
-mod notification_group;
 mod period;
+mod system_event;
 
+pub use client_event::*;
 pub use condition::*;
 pub use data_type::*;
-pub use events::*;
 pub use facilities::*;
 pub use notification::*;
-pub use notification_group::*;
 pub use period::*;
+pub use system_event::*;